Improve concurrency control
Now, at most two clients are waiting for a script at once, one that
is running it, and at most one waiting for the script to run. If
another client tries to execute the script, the one waiting on it
previously is "ejected" with no output, and the new client becomes the
"current waiting one".